# stock-python/contracts/openapi/public_api.yaml
openapi: 3.0.3
info:
  title: Stock API - Public
  description: Public API endpoints for stock trading platform
  version: 1.0.0
  contact:
    name: Stock API Support
servers:
  - url: /api/v1
    description: Public API v1
...
paths:
...
  /stocks/quote/{symbol}:
    get:
      summary: Get stock quote
      tags: [stocks]
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            default: yahoo
      responses:
        '200':
          description: Quote data
  /stocks/quotes/batch:
    post:
      summary: Get quotes for multiple symbols
      tags: [stocks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                symbols:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: List of quotes
  /stocks/search:
    get:
      summary: Search stocks
      tags: [stocks]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            default: yahoo
      responses:
        '200':
          description: Search results
  /stocks/history/{symbol}:
    get:
      summary: Get historical price data
      tags: [stocks]
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            default: 1mo
        - name: source
          in: query
          schema:
            type: string
            default: yahoo
      responses:
        '200':
          description: Historical data
  /stocks/watchlists:
    get:
      summary: Get user watchlists
      tags: [stocks]
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of watchlists
    post:
      summary: Create watchlist
      tags: [stocks]
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Watchlist created
  /stocks/watchlists/{watchlist_id}:
    get:
      summary: Get watchlist
      tags: [stocks]
      parameters:
        - name: watchlist_id
          in: path
          required: true
          schema:
            type: integer
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Watchlist data
    put:
      summary: Update watchlist
      tags: [stocks]
      parameters:
        - name: watchlist_id
          in: path
          required: true
          schema:
            type: integer
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Watchlist updated
    delete:
      summary: Delete watchlist
      tags: [stocks]
      parameters:
        - name: watchlist_id
          in: path
          required: true
          schema:
            type: integer
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Watchlist deleted
  /stocks/watchlists/{watchlist_id}/items:
    post:
      summary: Add stock to watchlist
      tags: [stocks]
      parameters:
        - name: watchlist_id
          in: path
          required: true
          schema:
            type: integer
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                symbol:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Stock added
  /stocks/watchlists/{watchlist_id}/items/{stock_id}:
    delete:
      summary: Remove stock from watchlist
      tags: [stocks]
      parameters:
        - name: watchlist_id
          in: path
          required: true
          schema:
            type: integer
        - name: stock_id
          in: path
          required: true
          schema:
            type: integer
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Stock removed

  # Portfolio
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...